// src/app/models/boveda.ts
export interface IBoveda {
  id?: string;
  fecha?: string;
  esactual?: string;
  idpeaje?: string;
  moneda001?: string;
  moneda005?: string;
  moneda01?: string;
  moneda025?: string;
  moneda05?: string;
  moneda1?: string;
  billete1?: string;
  billete2?: string;
  billete5?: string;
  billete10?: string;
  billete20?: string;
  observacion?: string;
  total?: string;
}